third_party/WebKit/LayoutTests/fast/images/image-document-zoom.html

<!DOCTYPE html>
<html>
<script>
var count = 0;

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.setCanOpenWindows();
    testRunner.waitUntilDone();
}

function log(message)
{
    var console = document.getElementById("console");
    console.appendChild(document.createTextNode(message + "\n"));
}

function imageSize(image)
{
    return { width: image.offsetWidth, height: image.offsetHeight };
}

function imageSizesEqual(size1, size2)
{
    var result = size1.width == size2.width && size1.height == size2.height;
    if (!result)
        log('SIZES NOT EQUAL: size1.width=' + size1.width + ', size1.height=' + size1.height + ', size2.width=' + size2.width + ', size2.height=' + size2.height);
    return result;
}

function zoomValue(view)
{
    return Math.round(view.devicePixelRatio * 100) + '%';
}


var originalImageSize;
var testScript = [ 'zoomPageOut', 'zoomPageIn', 'zoomPageIn', 'zoomPageOut', 'zoomPageOut', 'zoomPageOut', 'zoomPageIn', 'zoomPageIn'];

function testZoom(image) {

    originalImageSize = imageSize(image);

    testScript.forEach(function(action) {
        eventSender[action]();
        var sizeStayedTheSame = imageSizesEqual(originalImageSize, imageSize(image));
        log('Zoom at ' + zoomValue(image.ownerDocument.defaultView) + ' : ' + (sizeStayedTheSame ? 'PASS' : 'FAIL'));
    });

    testRunner.notifyDone();
}

var imageDocumentWindow = window.open("resources/dice.png", "test", "width=200,height=200");
imageDocumentWindow.onload = function() {
    if (!window.eventSender)
        return;

    var image = imageDocumentWindow.document.querySelector("img");
    if (image.offsetHeight == 0) {
      // On GTK+, sometimes the resize callback fires before the GTK
      // window has finished resizing. If that happens, try to resize
      // again.
      setTimeout(function() {
        testZoom(image);
      }, 0);
      return;
    }

    testZoom(image);
};

</script>
<body>
<p>This tests that page zoom and image auto-sizing interact well together.  This test requires testRunner to run.  To test manually, open <a href="resources/dice.png">this image</a> in a browser window, resize the window to 200px tall and zoom it in and out.  The image should get smaller or larger, respectively.</p>
<pre id="console"></pre>
</body>
</html>
