third_party/WebKit/LayoutTests/fast/images/image-document-zoom-expected.txt

This tests that page zoom and image auto-sizing interact well together. This test requires testRunner to run. To test manually, open this image in a browser window, resize the window to 200px tall and zoom it in and out. The image should get smaller or larger, respectively.

Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
Zoom at 100% : PASS
